batctl: Use command structure for remaining subcommands
The command datastructure was extended to be usable for mesh_iface and
non-mesh_iface subcommands. The referenced subcommands only have to use the
same function signature to work correctly.

The introduction of dummy parameter to the functions allows therefore to
convert all non-settings and non-debug-tables subcommands to the new
infrastructure.

Signed-off-by: Sven Eckelmann <sven@narfation.org>