batctl: Use external netlink socket for debug tables
The netlink socket is used by the debug table functionality to retrieve the
tables. The initialization can be centralized and moved to the main
function (controlled by a flag) to share it between the other commands.

Since more commands will use netlink in the future, this flag can reduce
the implementation effort significantly.

Signed-off-by: Sven Eckelmann <sven@narfation.org>