batctl: Add command to monitor for netlink events
The netlink multicast groups of batman-adv are currently used to inform
userspace about finished throughputmeter runs. It will later also be used
to inform about changes of settings of a specific batadv interface.

The "event" will wait for these messages and print brief summaries about
the received data. This can be used during the development of new features
or to monitor running systems.

Signed-off-by: Sven Eckelmann <sven@narfation.org>