batctl: Allow to enable/disable subcommands
Systems with limited space are usually not using all the functionality of
batman-adv and of batctl. Still, most of the batctl is still build even
when the corresponding batman-adv functionality would not be available.
This becomes especially important for subcommands which implement most
functionality in userspace when it was previously completely handled by the
kernel.

An external build system for such system can now just select the commands
it really needs. Combining this with link-time-optimizationa and/or
function/data-section garbage collection allows to find the right mix
between size and functionality.

Signed-off-by: Sven Eckelmann <sven@narfation.org>