batctl: Automatically translate vlan to mesh_iface
The way batctl is implementing support for VLANs is rather suboptimal. The
program has to guess whether it is started with an batadv or vlan interface
as argument.

Instead of distributing this over the whole program, do it on startup and
already retrieve the required mesh interface index and vid before the rest
of the program runs.

Signed-off-by: Sven Eckelmann <sven@narfation.org>